+++ lib/elixir_script/prelude/agent.ex
@@ -1,51 +1,53 @@
 defmodule ElixirScript.Agent do
-    @moduledoc false
+  @moduledoc false
+  require JS
 
   def start(fun, options \\ []) do
-    pid = spawn(fn() -> Process.sleep(:infinity) end)
+    pid = JS.new Elixir.Core.PID, []
 
-    if Elixir.Keyword.has_key?(options, :name) do
-      Process.register(pid, Elixir.Keyword.get(options, :name))
+    name = if Elixir.Keyword.has_key?(options, :name) do
+      Elixir.Keyword.get(options, :name)
+    else
+      nil
     end
 
-    Elixir.Core.Store.create(pid, fun.())
+    Elixir.Core.Store.create(pid, fun.(), name)
     { :ok, pid }
   end
 
   def start_link(fun, options \\ []) do
-    pid = spawn_link(fn() -> Process.sleep(:infinity) end)
+    pid = JS.new Elixir.Core.PID, []
 
-    if Elixir.Keyword.has_key?(options, :name) do
-      Process.register(pid, Elixir.Keyword.get(options, :name))
+    name = if Elixir.Keyword.has_key?(options, :name) do
+      Elixir.Keyword.get(options, :name)
+    else
+      nil
     end
 
-    Elixir.Core.Store.create(pid, fun.())
+    Elixir.Core.Store.create(pid, fun.(), name)
     { :ok, pid }
   end
 
   def stop(agent) do
-    Process.exit(agent)
+    Elixir.Core.Store.remove(agent)
     :ok
   end
 
   def update(agent, fun) do
-    pid = Elixir.Core.processes.pidof(agent)
-    current_state = Elixir.Core.Store.read(pid)
-    Elixir.Core.Store.update(pid, fun.(current_state))
+    current_state = Elixir.Core.Store.read(agent)
+    Elixir.Core.Store.update(agent, fun.(current_state))
     :ok
   end
 
   def get(agent, fun) do
-    pid = Elixir.Core.processes.pidof(agent)
-    current_state = Elixir.Core.Store.read(pid)
+    current_state = Elixir.Core.Store.read(agent)
     fun.(current_state)
   end
 
   def get_and_update(agent, fun) do
-    pid = Elixir.Core.processes.pidof(agent)
-    current_state = Elixir.Core.Store.read(pid)
+    current_state = Elixir.Core.Store.read(agent)
     {val, new_state} = fun.(current_state)
-    Elixir.Core.Store.update(pid, new_state)
+    Elixir.Core.Store.update(agent, new_state)
     val
   end