+++ src/javascript/lib/core/store.js
@@ -1,19 +1,42 @@
 let store = new Map();
+let names = new Map();
 
-function create(key, value){
+function get_key(key){
+  let real_key = key;
+
+  if(names.has(key)){
+    real_key = names.get(key);
+  }
+
+  if(store.has(real_key)){
+    return real_key
+  }
+
+  return new Error('Key Not Found');
+}
+
+function create(key, value, name = null){
+
+  if(name != null){
+    names.set(name, key);
+  }
+
   store.set(key, value);
 }
 
 function update(key, value){
-  store.set(key, value);
+  let real_key = get_key(key);
+  store.set(real_key, value);
 }
 
 function read(key){
-  return store.get(key);
+  let real_key = get_key(key);
+  return store.get(real_key);
 }
 
 function remove(key){
-  return store.delete(key);
+  let real_key = get_key(key);
+  return store.delete(real_key);
 }
 
 export default {